Move the sql query function to db module, remove offset and remove one returning parameter
